fix: bound hosted review wait loop
